Logical Operators in JavaScript
In JavaScript, there are several logical operators that can be used to combine and evaluate multiple expressions. These operators include:

The && operator compares two expressions and returns true only if both expressions are true. The || operator compares two expressions and returns true if either or both expressions are true. The ! operator negates the value of a single expression, so if the expression is true, it returns false and if the expression is false, it returns true.